requirements.txt
google-vizier  # Always compatible with most up-to-date version.
https://github.com/google-research/t5x/archive/master.zip  # Always compatible with most up to date version.
seqio  # Always compatible with most up-to-date version.

# Autodiff and numerical packages.
numpy==1.21.5
tensorflow>=2.9.1
jax==0.3.16
jaxlib==0.3.15
flax==0.6.0

# Standard packages.
typing  # Version dependent on Python version.
pytest  # Use the latest version to match github workflow.
absl-py>=0.7,<1.0

# Tidy
ml-collections==0.1.1

# Proto
protobuf>=3.6,<4.0

